requirements.txt
pybatsim==3.2.1
numpy==1.21.5
progressbar2==3.53.1
pandas==1.3.5
evalys==4.0.6
keras==2.12.0